♻️ refactor: 에러 메시지 위치 변경, 카테고리 엔티티 삭제 컬럼 데코레이터 변경
